+++ rotkehlchen/chain/ethereum/modules/oneinch/v3/decoder.py
@@ -1,20 +1,19 @@
 from typing import TYPE_CHECKING
 
-from rotkehlchen.chain.ethereum.modules.oneinch.constants import CPT_ONEINCH_V5
 from rotkehlchen.chain.evm.decoding.oneinch.decoder import OneinchCommonDecoder
 from rotkehlchen.chain.evm.decoding.oneinch.v4.decoder import Oneinchv3n4DecoderBase
 from rotkehlchen.chain.evm.decoding.types import CounterpartyDetails
-from rotkehlchen.chain.evm.types import string_to_evm_address
 
 if TYPE_CHECKING:
-    from rotkehlchen.chain.evm.decoding.base import BaseDecoderTools
     from rotkehlchen.chain.evm.node_inquirer import EvmNodeInquirer
     from rotkehlchen.user_messages import MessagesAggregator
+    from rotkehlchen.chain.evm.decoding.base import BaseDecoderTools
 
-ONEINCH_V5_ROUTER = string_to_evm_address('0x1111111254EEB25477B68fb85Ed929f73A960582')
+from ..constants import CPT_ONEINCH_V3
+from .constants import ONEINCH_V3_MAINNET_ROUTER
 
 
-class Oneinchv5Decoder(Oneinchv3n4DecoderBase):
+class Oneinchv3Decoder(Oneinchv3n4DecoderBase):
 
     def __init__(
             self,
@@ -26,12 +25,12 @@ class Oneinchv5Decoder(Oneinchv3n4DecoderBase):
             evm_inquirer=evm_inquirer,
             base_tools=base_tools,
             msg_aggregator=msg_aggregator,
-            router_address=ONEINCH_V5_ROUTER,
-            counterparty=CPT_ONEINCH_V5,
+            router_address=ONEINCH_V3_MAINNET_ROUTER,
+            counterparty=CPT_ONEINCH_V3,
         )
 
     # -- DecoderInterface methods
 
     @staticmethod
     def counterparties() -> tuple[CounterpartyDetails, ...]:
-        return OneinchCommonDecoder.generate_counterparty_details(CPT_ONEINCH_V5)
+        return OneinchCommonDecoder.generate_counterparty_details(CPT_ONEINCH_V3)