+++ rotkehlchen/chain/ethereum/modules/oneinch/v4/decoder.py
@@ -2,7 +2,7 @@ from typing import TYPE_CHECKING
 
 from rotkehlchen.chain.ethereum.modules.oneinch.constants import CPT_ONEINCH_V4
 from rotkehlchen.chain.evm.decoding.oneinch.v4.constants import ONEINCH_V4_ROUTER
-from rotkehlchen.chain.evm.decoding.oneinch.v4.decoder import Oneinchv4DecoderBase
+from rotkehlchen.chain.evm.decoding.oneinch.v4.decoder import Oneinchv3n4DecoderBase
 
 if TYPE_CHECKING:
     from rotkehlchen.chain.evm.decoding.base import BaseDecoderTools
@@ -10,7 +10,7 @@ if TYPE_CHECKING:
     from rotkehlchen.user_messages import MessagesAggregator
 
 
-class Oneinchv4Decoder(Oneinchv4DecoderBase):
+class Oneinchv4Decoder(Oneinchv3n4DecoderBase):
 
     def __init__(
             self,